[front/observability] - refactor: streamline SourceChart and extract color utils
- Moved color-related constants and `getSourceColor` function from `SourceChart` to `constants` to improve code reusability
- Defined `SourceChartDatum` type for cleaner data handling within `SourceChart`
- Removed unnecessary checks and simplified data fetching for contextOrigin in `SourceChart` component
- Cleaned up `SourceChart` props definition for better readability and maintenance
