+++ front/components/agent_builder/observability/constants.ts
@@ -49,3 +49,16 @@ export const FEEDBACK_DISTRIBUTION_LEGEND = [
   { key: "positive", label: "Positive" },
   { key: "negative", label: "Negative" },
 ] as const;
+
+const SOURCE_COLORS = [
+  "text-pink-300 dark:text-pink-300-night",
+  "text-blue-400 dark:text-blue-400-night",
+  "text-rose-400 dark:text-rose-400-night",
+  "text-amber-400 dark:text-amber-400-night",
+  "text-violet-300 dark:text-violet-300-night",
+  "text-slate-300 dark:text-slate-300-night",
+] as const;
+
+export function getSourceColor(index: number) {
+  return SOURCE_COLORS[index % SOURCE_COLORS.length];
+}
